requirements.txt
fastapi==0.114.2
uvicorn==0.34.2
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.27.2
pydantic==2.11.3
python-multipart==0.0.20
r2r==3.4.1
typing-extensions==4.9.0
openai==1.12.0
python-jose==3.3.0 